freiland/theme: only show events in data archives
this filters posts in the categegory configured in ec3
when showing date archives.
Only does this when ec3 is installed/activated.
This has the advantage of showing pretty urls when selecting a date
in the event calendar in contrast to the urls produced by ec3 when
using its internal 'only show event posts' option.