ec3: always only show event posts
This bypasses the ec3 option 'Show all categories in calendar'
This is needed by the previous commit to show pretty urls when
using the event calendar and only show posts in the event category.

This hacks the ec3 sql generating code
to always only show posts in the events category.

So this might break when upgrading ec3.