smwimport: create new link for each new attachment
this keeps the 1 to 1 relation between attachments and files
and prevents broken links when an attachment is deleted and
another attachment has the same file

Also only create a link if the attachment does not yet exist.
This prevents having multiple links to the same file in
different monthly upload folders.